Updated miscs.test.js to use entityTestFactory
